+++ R/2-mod.R
@@ -45,6 +45,10 @@ modHistory <- function(object, reference){
         packageCheck("balance")
         exprs.i <- t(balance::balance.fromSBP(data, indexedModel))
         colnames(exprs.i) <- rownames(data)
+      }else if("rda" %in% class(indexedModel)){
+        packageCheck("vegan")
+        exprs.i <- t(predict(indexedModel, data, type = "wa"))
+        colnames(exprs.i) <- rownames(data)
       }else{
         stop("Reduction model not recognized.")
       }
@@ -258,7 +262,7 @@ modCLR <- function(object){
              "This function is applied to the results of ?exprso.")
 
   logX <- log(object@exprs)
-  object@exprs <- apply(logX, 2, function(x) x / mean(x))
+  object@exprs <- apply(logX, 2, function(x) x - mean(x))
   return(object)
 }