+++ tests/testthat/test-fs-etc.R
@@ -35,16 +35,6 @@ if(requireNamespace("limma", quietly = TRUE) &
       rownames(fsMrmre(array, top = 0)@exprs),
       c("feat4", "feat2", "feat1", "feat3")
     )
-
-    expect_equal(
-      rownames(fsPathClassRFE(array, top = 0)@exprs),
-      c("feat2", "feat4")
-    )
-
-    expect_equal(
-      rownames(fsPropd(array, top = 0)@exprs),
-      c("feat1", "feat4", "feat2", "feat3")
-    )
   })
 }